+++ pac4j-core/src/main/java/org/pac4j/core/client/direct/AnonymousClient.java
@@ -7,7 +7,7 @@ import org.pac4j.core.profile.AnonymousProfile;
 import java.util.Optional;
 
 /**
- * Anonymous client. Not to be used except for advanced use cases.
+ * Anonymous client.
  *
  * @author Jerome Leleu
  * @since 1.8.1
@@ -20,7 +20,8 @@ public final class AnonymousClient extends DirectClient<AnonymousCredentials> {
 
     public AnonymousClient() {
         if (!warned) {
-            logger.warn("AnonymousClient is an advanced feature: be careful when using it to avoid any security issues!");
+            logger.warn("Be careful when using the 'AnonymousClient': an 'AnonymousProfile' is returned "
+                + "and the access is granted for the request.");
             warned = true;
         }
     }
@@ -28,8 +29,6 @@ public final class AnonymousClient extends DirectClient<AnonymousCredentials> {
     @Override
     protected void clientInit() {
         defaultCredentialsExtractor(ctx -> Optional.of(AnonymousCredentials.INSTANCE));
-        defaultAuthenticator((cred, ctx )-> {
-            cred.setUserProfile(AnonymousProfile.INSTANCE);
-        });
+        defaultAuthenticator((cred, ctx )-> cred.setUserProfile(AnonymousProfile.INSTANCE));
     }
 }