+++ pac4j-core/src/test/java/org/pac4j/core/client/direct/AnonymousClientTest.java
@@ -1,0 +1,15 @@
+package org.pac4j.core.client.direct;
+
+import org.junit.Test;
+
+/**
+ * @author Jerome LELEU
+ * @since 4.0.1
+ */
+public class AnonymousClientTest {
+
+    @Test
+    public void test() {
+        new AnonymousClient();
+    }
+}